State: Move getJetpackSettings to global selectors directory
