+++ client/state/jetpack/settings/test/selectors.js
@@ -4,7 +4,6 @@
 import { expect } from 'chai';
 
 import {
-	getJetpackSettings,
 	getJetpackSetting,
 	getJetpackSettingsSaveRequestStatus,
 	isJetpackSettingsSaveSuccessful,
@@ -14,36 +13,6 @@ import {
 import { settings as SETTINGS_FIXTURE } from './fixture';
 
 describe( 'selectors', () => {
-	describe( '#getJetpackSettings', () => {
-		it( 'should return settings for all modules for a known site', () => {
-			const stateIn = {
-					jetpack: {
-						settings: {
-							items: SETTINGS_FIXTURE
-						}
-					}
-				},
-				siteId = 12345678;
-			const output = getJetpackSettings( stateIn, siteId );
-			expect( output ).to.eql( SETTINGS_FIXTURE[ siteId ] );
-		} );
-
-		it( 'should return null for an unknown site', () => {
-			const stateIn = {
-					jetpack: {
-						settings: {
-							items: {
-								654321: SETTINGS_FIXTURE
-							}
-						}
-					}
-				},
-				siteId = 12345678;
-			const output = getJetpackSettings( stateIn, siteId );
-			expect( output ).to.be.null;
-		} );
-	} );
-
 	describe( '#getJetpackSetting', () => {
 		it( 'should return a certain setting for a known site', () => {
 			const stateIn = {
